Refactor out timerState enum to its own file
