+++ lib/widgets/control_row/buttons/pause_or_resume_button.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:orodomop/models/timer_model.dart';
+import 'package:orodomop/models/timer_state.dart';
 import 'package:provider/provider.dart';
 
 class PauseOrResumeButton extends StatelessWidget {
